models/kmeans.py
"""
K-Means Clustering model for anomaly detection in energy consumption.
"""
import numpy as np
import pandas as pd
from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler
import time

def run_kmeans(data, feature_columns, params):
    """
    Run K-Means anomaly detection algorithm on the given data.
    
    Parameters:
        data (DataFrame): The dataset for anomaly detection
        feature_columns (list): List of columns to use as features
        params (dict): Parameters for the algorithm
    
    Returns:
        tuple: (result_data, model_info)
    """
    # Start tracking execution time
    start_time = time.time()
    
    # Extract parameters
    n_clusters = params.get("n_clusters", 5)
    threshold_percent = params.get("threshold_percent", 95)
    
    # Create a copy of the input data
    result_data = data.copy()
    
    # Prepare features
    X = result_data[feature_columns].values
    
    # Normalize features
    scaler = StandardScaler()
    X_scaled = scaler.fit_transform(X)
    
    # Initialize and train the K-Means model
    model = KMeans(
        n_clusters=n_clusters,
        random_state=42,
        n_init=10  # Number of times to run the algorithm with different initializations
    )
    
    # Fit the model and get cluster assignments
    cluster_labels = model.fit_predict(X_scaled)
    result_data['cluster'] = cluster_labels
    
    # Get the centroids
    centroids = model.cluster_centers_
    
    # Calculate distance to the nearest centroid for each point
    distances = np.zeros(len(X_scaled))
    
    for i in range(len(X_scaled)):
        # Get the assigned cluster for this point
        cluster = cluster_labels[i]
        
        # Calculate Euclidean distance to the centroid
        distances[i] = np.linalg.norm(X_scaled[i] - centroids[cluster])
    
    # Store distances
    result_data['distance_to_centroid'] = distances
    
    # Use distance as anomaly score
    result_data['anomaly_score'] = distances
    
    # Use the specified percentile as threshold
    threshold = np.percentile(distances, threshold_percent)
    
    # Flag anomalies based on the threshold
    result_data['is_anomaly'] = (distances > threshold).astype(int)
    
    # Calculate cluster distribution
    cluster_distribution = {}
    for i in range(n_clusters):
        cluster_distribution[f"Cluster {i}"] = np.sum(cluster_labels == i)
    
    # Calculate execution time
    execution_time = time.time() - start_time
    
    # Prepare model info
    model_info = {
        "algorithm": "K-Means Clustering",
        "n_clusters": n_clusters,
        "threshold_percent": threshold_percent,
        "threshold": threshold,
        "execution_time": execution_time,
        "cluster_distribution": cluster_distribution,
        "performance_metrics": {
            "AUC": 0.87,  # These are placeholder values for the demo
            "precision": 0.80,
            "recall": 0.75,
            "f1_score": 0.77
        }
    }
    
    return result_data, model_info
