pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
    "streamlit-extras>=0.6.0",
    "streamlit>=1.44.1",
    "seaborn>=0.13.2",
    "pandas>=2.2.3",
    "numpy>=2.2.5",
    "plotly>=6.0.1",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.14.0",
]
